Added files via upload
The care.css is the one that I made. I also made changes to the styles.css to improve it style. In the styles.css some of the thing will have a either a modified or added comment next to them. Those are the one that I made change to.